# docs/next-project/openapi.yaml
openapi: 3.1.0
info:
  title: MES Agent Hub API (초안)
  version: 0.1.0-draft
  description: |
    새 저장소 서버 API 초안. 데모의 저장소 함수(src/db/repositories/*)와 1:1로 대응한다(docs/architecture/data-contract.md §6).
    인증은 사내 SSO 세션 쿠키. 역할: member(담당자), system_owner, requester.
    오류는 공통 Problem 형식. 409는 규칙 위반(진행 중 요청, 참조 중 삭제, 완료된 대화 변경 등).
servers:
  - url: /api
security:
  - session: []
...
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: 회원가입 (local 모드) — ID·비밀번호만, 가입 즉시 세션 쿠키
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Credentials' } } } }
      responses:
        '201': { description: 가입·로그인됨, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { description: ID·비밀번호 규칙 위반 }
        '404': { description: local 모드가 아님 }
        '409': { description: 이미 사용 중인 ID }
  /auth/login:
    post:
      tags: [auth]
      summary: 로그인 (local 모드) — 세션 쿠키 발급
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Credentials' } } } }
      responses:
        '200': { description: 로그인됨, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { description: ID 또는 비밀번호 불일치 (구분하지 않음) }
        '404': { description: local 모드가 아님 }
    get:
      tags: [auth]
      summary: SSO 로그인 시작 (oidc 모드, IdP로 리다이렉트)
      security: []
      responses:
        '302': { description: IdP로 이동 }
        '404': { description: oidc 모드가 아님 }
...
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: hub_session }
...
  schemas:
    Credentials:
      type: object
      required: [loginId, password]
      properties:
        loginId: { type: string, pattern: '^[a-z0-9._-]{3,32}$', description: 로그인 ID = 계정 식별자 (D34) }
        password: { type: string, minLength: 8, maxLength: 128 }
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        roles: { type: array, items: { enum: [member, system_owner, requester] } }
...